requirements.txt
fastapi 
uvicorn 
pydantic
google-cloud-firestore
dotenv